first minor restructuring of the modules and classes
